requirements.txt
keyboard==0.13.5
hidapi==0.10.1
PyYAML==5.4.1
pywinauto==0.6.6
watchdog==2.1.6